requirements.txt
alembic==1.0.11
aniso8601==7.0.0
beautifulsoup4==4.8.0
blinker==1.4
Click==7.0
Flask==1.0.3
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Limiter==1.0.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
limits==1.3
lxml==4.4.1
Mako==1.0.13
MarkupSafe==1.1.1
Pillow==6.1.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
redis==3.3.5
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.5
Werkzeug==0.15.4
